fix subnautica like icon problem
when imported from the asset libary the icon would be missing since it is in the project root.
But the subnautica like reporter would crash if there was no icon.png. Standart for Godot4 is icon.svg
icon was moved to the addon folder where it belongs
and it now uses the icon path from project settings.